Move the chat_service crate to server::services::chat
