+++ Cargo.toml
@@ -2,7 +2,6 @@
 members = [
     "server",
     "chatroom",
-    "chat-service",
 ]
 
 resolver = "2"
