--- chat-service/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "chat-service"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-chatroom = { version = "0.1.0", path = "../chatroom" }
-axum = {workspace = true}
-tokio = {workspace = true}
-serde = {workspace = true, features = ["rc"]}
-tower-http = {workspace = true}
-http = {workspace = true}
-
-# [dev-dependencies]
-# axum-test = "15"
