+++ server/Cargo.toml
@@ -5,7 +5,6 @@ edition = "2021"
 
 [dependencies]
 chatroom = { version = "0.1.0", path = "../chatroom" }
-chat-service = { version = "0.1.0", path = "../chat-service" }
 
 axum = {workspace = true}
 # tower = {version = "0.4", features = ["full"]}
@@ -18,7 +17,7 @@ tokio = {workspace = true}
 
 # chrono = "0.4"
 
-serde = {workspace = true}
+serde = {workspace = true, features = ["rc"]}
 # serde_json = "1"
 toml = "0.8"
